new_implementation/requirements.txt
# Core web framework and API
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.11.0,<3.0.0

# Database and ORM
sqlalchemy>=2.0.40,<3.0.0
psycopg2-binary>=2.9.10,<3.0.0
alembic>=1.16.0,<2.0.0

# HTTP requests and async
requests>=2.31.0,<3.0.0
httpx>=0.28.0,<1.0.0

# Telegram bot
python-telegram-bot>=22.0,<23.0

# Image processing and rendering
Pillow>=11.0.0,<13.0.0
cairosvg>=2.7.0,<3.0.0
numpy>=2.0.0,<3.0.0

# Browser automation (for map rendering)
selenium>=4.34.0,<5.0.0

# Testing framework
pytest>=8.4.0,<9.0.0
pytest-asyncio>=1.1.0,<2.0.0
pytest-cov>=7.0.0,<8.0.0
pytest-mock>=3.12.0,<4.0.0
coverage>=7.11.0,<8.0.0

# Timezone handling
pytz>=2024.2

# Additional utilities
python-multipart>=0.0.17  # For FastAPI file uploads
python-dotenv>=1.0.1     # For environment variable management
click>=8.1.0,<9.0.0      # For command-line interface